fix ml pipeline index

+++ plugins/dbnd-airflow-operator/test_dbnd_airflow_operator/airflow_home/dags/dag_with_ml_pipeline.py
@@ -26,7 +26,7 @@ default_args = {
 
 
 def _split(data):
-    half_len = len(data) / 2
+    half_len = int(len(data) / 2)
     return data.iloc[:, :half_len], data.iloc[:, half_len:]
